Refactoring: shared *.h files moved to include directory(to be included to SDK in future), include dir set as -I in compiler FLAGS, fixed some errors in datanase module, fixed some errors in rtcp module.
