+++ include/captagent/globals.h
