+++ modules.am
@@ -1,7 +1,7 @@
 moddir = $(libdir)/captagent/modules
 
-AM_CPPFLAGS = ${AM_INCLUDES}
-AM_CFLAGS = ${AM_CFLAGS}
+AM_CPPFLAGS = ${AM_INCLUDES} -I$(top_srcdir)/include
+AM_CFLAGS = ${AM_CFLAGS} -I$(top_srcdir)/include
 
 CLEANFILES = *.la *.lo *.o *.so *.slo .*
 DISTCLEANFILES = Makefile.in
