+++ src/Makefile.am
@@ -1,9 +1,21 @@
-include ../common.am
+include $(top_srcdir)/common.am
+
+SUBDIRS = \
+	. \
+	modules/socket/pcap \
+	modules/socket/raw \
+	modules/socket/rtcpxr \
+	modules/protocol/sip \
+	modules/protocol/rtcp \
+	modules/transport/hep \
+	modules/transport/json \
+	modules/database/hash
+
 
 bin_PROGRAMS = captagent
 BUILT_SOURCES = capplan.tab.h
-AM_CFLAGS = -g -fPIC -rdynamic
-AM_CPPFLAGS = -DSYSCONFDIR='"$(sysconfdir)"'
+AM_CFLAGS = -g -fPIC -rdynamic -I$(top_srcdir)/include
+AM_CPPFLAGS = -DSYSCONFDIR='"$(sysconfdir)"' -I$(top_srcdir)/include
 captagent_SOURCES = captagent.c xmlread.c modules.c log.c md5.c conf_function.c capplan.tab.c lex.yy.c
 captagent_LDADD = ${PTHREAD_LIBS} ${EXPAT_LIBS} ${DL_LIBS} ${FLEX_LIBS}
 config = $(sysconfdir)/$(bin_PROGRAMS)/$(bin_PROGRAMS).xml
