+++ src/modules/database/hash/Makefile.am
@@ -1,7 +1,9 @@
-include ../../../../modules.am
+include $(top_srcdir)/modules.am
+
+SUBDIRS = .
 #
 database_hash_la_SOURCES = database_hash.c captarray.c localapi.c
-database_hash_la_CFLAGS = -Wall ${MODULE_CFLAGS} ${AM_CFLAGS} ${EXPAT_LIBS}
+database_hash_la_CFLAGS = -Wall ${MODULE_CFLAGS} ${EXPAT_LIBS}
 database_hash_la_LDFLAGS = -module -avoid-version
 
 include_HEADERS =
