+++ src/modules/database/hash/database_hash.h
@@ -27,7 +27,7 @@
 #define DATABASE_HASH_H_
 
 
-#include "../../xmlread.h"
+#include <captagent/xmlread.h>
 #include "uthash.h"
 #include "hash_structure.h"
