+++ src/modules/database/hash/localapi.c
@@ -24,11 +24,11 @@
 */
 
 #include <stdio.h>
-#include "src/api.h"
-#include "src/structure.h"
-#include "src/modules_api.h"
-#include "src/modules.h"
-#include "src/log.h"
+#include <captagent/api.h>
+#include <captagent/structure.h>
+#include <captagent/modules_api.h>
+#include <captagent/modules.h>
+#include <captagent/log.h>
 #include "localapi.h"
 
 char* hashapi_lookup(char *ip, int port)
