+++ src/modules/protocol/rtcp/Makefile.am
@@ -1,4 +1,6 @@
-include ../../../../modules.am
+include $(top_srcdir)/modules.am
+
+SUBDIRS = .
 #
 protocol_rtcp_la_SOURCES = protocol_rtcp.c parser_rtcp.c
 protocol_rtcp_la_CFLAGS = -Wall ${MODULE_CFLAGS}
