+++ src/modules/protocol/rtcp/parser_rtcp.c
@@ -24,7 +24,7 @@
 */
 
 #include <stdio.h>
-#include "src/log.h"
+#include <captagent/log.h>
 #include "parser_rtcp.h"
