+++ src/modules/protocol/rtcp/protocol_rtcp.h
@@ -26,7 +26,7 @@
 #ifndef _PROTOCOL_SIP_H_
 #define _PROTOCOL_SIP_H_
 
-#include "../../xmlread.h"
+#include <captagent/xmlread.h>
 #include "parser_rtcp.h"
