+++ src/modules/protocol/sip/Makefile.am
@@ -1,4 +1,6 @@
-include ../../../../modules.am
+include $(top_srcdir)/modules.am
+
+SUBDIRS = .
 #
 protocol_sip_la_SOURCES = protocol_sip.c parser_sip.c
 protocol_sip_la_CFLAGS = -Wall ${MODULE_CFLAGS}
