+++ src/modules/protocol/sip/protocol_sip.c
@@ -29,12 +29,12 @@
 #include <stdint.h>
 #include <inttypes.h>
 
-#include "src/api.h"
-#include "src/structure.h"
-#include "src/modules_api.h"
-#include "src/modules.h"
+#include <captagent/api.h>
+#include <captagent/structure.h>
+#include <captagent/modules_api.h>
+#include <captagent/modules.h>
 #include "protocol_sip.h"
-#include "src/log.h"
+#include <captagent/log.h>
 
 xml_node *module_xml_config = NULL;
 char *module_name="protocol_sip";
