+++ src/modules/socket/pcap/Makefile.am
@@ -1,4 +1,6 @@
-include ../../../../modules.am
+include $(top_srcdir)/modules.am
+
+SUBDIRS = .
 #
 socket_pcap_la_SOURCES = socket_pcap.c ipreasm.c
 socket_pcap_la_CFLAGS = -Wall ${MODULE_CFLAGS} ${LUA_CFLAGS}
