+++ src/modules/socket/pcap/socket_pcap.c
@@ -56,15 +56,15 @@
 
 #include <pcap.h>
 
-#include "src/capture.h"
-#include "src/globals.h"
-#include "src/api.h"
-#include "src/proto_sip.h"
-#include "src/structure.h"
-#include "src/modules_api.h"
-#include "src/modules.h"
+#include <captagent/capture.h>
+#include <captagent/globals.h>
+#include <captagent/api.h>
+#include <captagent/proto_sip.h>
+#include <captagent/structure.h>
+#include <captagent/modules_api.h>
+#include <captagent/modules.h>
 #include "socket_pcap.h"
-#include "src/log.h"
+#include <captagent/log.h>
 #include "ipreasm.h"
 
 xml_node *module_xml_config = NULL;
