+++ src/modules/socket/pcap/socket_pcap.h
@@ -27,7 +27,7 @@
 #ifndef _SOCKET_PCAP_H_
 #define _SOCKET_PCAP_H_
 
-#include "../../xmlread.h"
+#include <captagent/xmlread.h>
 
 #define FILTER_LEN 4080
