+++ src/modules/socket/raw/Makefile.am
@@ -1,4 +1,6 @@
-include ../../../../modules.am
+include $(top_srcdir)/modules.am
+
+SUBDIRS = .
 #
 socket_raw_la_SOURCES = socket_raw.c localapi.c
 socket_raw_la_CFLAGS = -Wall ${MODULE_CFLAGS} ${LUA_CFLAGS}
