+++ src/modules/socket/raw/socket_raw.c
@@ -59,15 +59,15 @@
 
 #include <pcap.h>
 
-#include "src/capture.h"
-#include "src/globals.h"
-#include "src/api.h"
-#include "src/proto_sip.h"
-#include "src/structure.h"
-#include "src/modules_api.h"
-#include "src/modules.h"
+#include <captagent/capture.h>
+#include <captagent/globals.h>
+#include <captagent/api.h>
+#include <captagent/proto_sip.h>
+#include <captagent/structure.h>
+#include <captagent/modules_api.h>
+#include <captagent/modules.h>
 #include "socket_raw.h"
-#include "src/log.h"
+#include <captagent/log.h>
 #include "localapi.h"
