+++ src/modules/socket/raw/socket_raw.h
@@ -27,7 +27,7 @@
 #ifndef _socket_raw_H_
 #define _socket_raw_H_
 
-#include "../../xmlread.h"
+#include <captagent/xmlread.h>
 
 #define FILTER_LEN 4080
