+++ src/modules/socket/rtcpxr/Makefile.am
@@ -1,4 +1,6 @@
-include ../../../../modules.am
+include $(top_srcdir)/modules.am
+
+SUBDIRS = .
 #
 socket_rtcpxr_la_SOURCES = socket_rtcpxr.c
 socket_rtcpxr_la_CFLAGS = -Wall ${MODULE_CFLAGS}
